conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_sysroot_linux-64_curr_repodata_hack=3=h69a702a_16
absl-py=2.1.0=pypi_0
adjusttext=1.2.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.2=pypi_0
annotated-types=0.6.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
binutils_impl_linux-64=2.38=h2a08ee3_1
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.7.4=hbcca054_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chex=0.1.86=pypi_0
click=8.1.7=pypi_0
cloudpickle=2.2.1=py311h06a4308_0
colorama=0.4.6=pypi_0
colorlog=6.8.2=pypi_0
colourmap=1.1.17=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
copier=9.2.0=pypi_0
cycler=0.12.1=pypi_0
dataproperty=1.0.1=pypi_0
datasets=2.21.0=pypi_0
datazets=0.1.9=pypi_0
debugpy=1.8.5=py311hf86e51f_0
decorator=4.4.2=pypi_0
dill=0.3.8=pypi_0
dm-control=1.0.14=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
docstring-parser=0.16=pypi_0
dunamai=1.21.0=pypi_0
etils=1.8.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
farama-notifications=0.0.4=py311h06a4308_0
ffmpeg=1.4=pypi_0
filelock=3.13.4=pypi_0
flax=0.8.4=pypi_0
fonttools=4.51.0=pypi_0
freetype-py=2.4.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
funcy=2.0=pypi_0
gcc=12.1.0=h9ea6d83_10
gcc_impl_linux-64=12.1.0=hea43390_17
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glfw=2.7.0=pypi_0
google-api-core=2.19.1=pypi_0
google-auth=2.32.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-storage=2.17.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.7.1=pypi_0
googleapis-common-protos=1.63.2=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.2=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.29.1=pypi_0
gymnasium-robotics=1.2.4=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.23.4=pypi_0
huggingface-sb3=3.0=pypi_0
idna=3.7=pypi_0
imageio=2.34.1=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
imitation=1.0.1.dev176883612=pypi_0
importlib-metadata=8.4.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=8.4.0=hd8ed1ab_0
iniconfig=2.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
jax=0.4.28=pypi_0
jax-cuda12-pjrt=0.4.28=pypi_0
jax-cuda12-plugin=0.4.28=pypi_0
jax-jumpy=1.0.0=py311h06a4308_0
jaxlib=0.4.28=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jinja2-ansible-filters=1.3.2=pypi_0
joblib=1.4.2=pypi_0
jsonpickle=3.2.2=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py311h38be061_0
kernel-headers_linux-64=3.10.0=h4a8ded7_16
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.3=h659f571_0
labmaze=1.0.6=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libedit=3.1.20191231=he28a2e2_2
libffi=3.4.4=h6a678d5_0
libgcc-devel_linux-64=12.1.0=h1ec3361_17
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libsanitizer=12.1.0=ha89aaad_17
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=14.1.0=hc0a3c3a_0
libuuid=1.41.5=h5eee18b_0
loco-mujoco=0.3.0=pypi_0
lxml=5.2.1=pypi_0
mako=1.3.5=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
matplotlib2tikz=0.7.6=pypi_0
mbstrdecoder=1.1.3=pypi_0
mdurl=0.1.2=pypi_0
minari=0.4.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
ml-dtypes=0.4.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
mujoco=2.3.7=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
munch=4.0.0=pypi_0
mushroom-rl=1.10.1=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pypi_0
num2words=0.5.13=pypi_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvcc-cu12=12.5.40=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
open-loop-locomotion=0.0.0=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.3.1=h4bc722e_2
opt-einsum=3.3.0=pypi_0
optax=0.2.2=pypi_0
optuna=3.6.1=pypi_0
orbax-checkpoint=0.5.15=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pathspec=0.12.1=pypi_0
pathvalidate=3.2.0=pypi_0
patsy=0.5.6=pypi_0
pca=2.0.7=pypi_0
pettingzoo=1.24.3=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=pypi_0
pip=24.0=pypi_0
platformdirs=4.2.2=pyhd8ed1ab_0
plotly=5.22.0=pypi_0
pluggy=1.5.0=pypi_0
plumbum=1.8.3=pypi_0
portion=2.4.2=pypi_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.36=pypi_0
proto-plus=1.24.0=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=16.1.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pygame=2.6.0=pypi_0
pyglet=2.0.16=pypi_0
pygments=2.17.2=pypi_0
pyopengl=3.1.0=pypi_0
pyparsing=3.1.2=pypi_0
pyrender=0.1.45=pypi_0
pytablewriter=1.2.0=pypi_0
pytest=8.3.2=pypi_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.11=2_cp311
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.1.1=py311h759c1eb_0
questionary=2.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rl-zoo3=2.3.0=pypi_0
rsa=4.9=pypi_0
sacred=0.8.5=pypi_0
sb3-contrib=2.3.0=pypi_0
scatterd=1.3.7=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
seaborn=0.13.2=pypi_0
seals=0.2.2.dev152004815+d20240910=pypi_0
sentry-sdk=2.0.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py311h06a4308_0
shellingham=1.5.4=pypi_0
shtab=1.7.1=pypi_0
six=1.16.0=pyh6c4a22f_0
skrl=1.2.0=pypi_0
smmap=5.0.1=pypi_0
snake-env=0.0.1=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlalchemy=2.0.31=pypi_0
sqlite=3.41.2=h5eee18b_0
stable-baselines3=2.3.2=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.2=pypi_0
sympy=1.12=pypi_0
sysroot_linux-64=2.17=h4a8ded7_16
tabledata=1.3.3=pypi_0
tbb=2021.8.0=hdb19cb5_0
tcolorpy=0.1.6=pypi_0
tenacity=8.3.0=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorstore=0.1.60=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.1=pypi_0
torch=2.3.1=pypi_0
torch-geometric=2.5.3=pypi_0
torchvision=0.18.1=pypi_0
tornado=6.4.1=py311h331c9d8_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
trimesh=4.4.3=pypi_0
triton=2.3.1=pypi_0
typepy=1.3.2=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.9.0=py311h06a4308_1
typing_extensions=4.9.0=py311h06a4308_1
tyro=0.8.3=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wandb=0.16.6=pypi_0
wasabi=1.1.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.2=pypi_0
wget=3.2=pypi_0
wheel=0.41.2=py311h06a4308_0
wrapt=1.16.0=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h75354e8_4
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
